+++ python/pyspark/sql/tests/pandas/streaming/test_transform_with_state_state_variable_checkpoint_v2.py
@@ -1,0 +1,52 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import unittest
+
+from pyspark.testing.sqlutils import ReusedSQLTestCase
+from pyspark.sql.tests.pandas.streaming.test_transform_with_state_state_variable import (
+    TransformWithStateInPySparkStateVariableTestsMixin,
+)
+
+
+class TransformWithStateInPySparkStateVariableWithCheckpointV2TestsMixin(
+    TransformWithStateInPySparkStateVariableTestsMixin
+):
+    @classmethod
+    def conf(cls):
+        cfg = super().conf()
+        cfg.set("spark.sql.streaming.stateStore.checkpointFormatVersion", "2")
+        return cfg
+
+
+class TransformWithStateInPySparkStateVariableWithCheckpointV2Tests(
+    TransformWithStateInPySparkStateVariableWithCheckpointV2TestsMixin,
+    ReusedSQLTestCase
+):
+    pass
+
+
+if __name__ == "__main__":
+    from pyspark.sql.tests.pandas.streaming.test_transform_with_state_state_variable_checkpoint_v2 import *  # noqa: F401,E501
+
+    try:
+        import xmlrunner
+
+        testRunner = xmlrunner.XMLTestRunner(output="target/test-reports", verbosity=2)
+    except ImportError:
+        testRunner = None
+    unittest.main(testRunner=testRunner, verbosity=2)
